* Info widgets use alternate color for info by default * Toolbars that close their child windows because the button for them is   disabled will reopen those windows when the button is enabled again * Added a separator at the top of the ship cargo list * Cargo sychronized with remote clients (Issue 128) * Can send N_send() arguments through individual network calls now; use   N_send_start(), followed by N_send_*(), and finally N_send(client, NULL) * N_connect() resolves hostnames and parses port

+++ src/game/g_shared.h
@@ -70,8 +70,8 @@ typedef struct g_nation {
 
 /* Cargo and trading settings structure */
 typedef struct g_cargo {
-        int capacity, amounts[G_CARGO_TYPES],
-            prices[G_CARGO_TYPES], maxs[G_CARGO_TYPES];
+        short capacity, amounts[G_CARGO_TYPES],
+              prices[G_CARGO_TYPES], maxs[G_CARGO_TYPES];
 } g_cargo_t;
 
 /* g_client.c */
